// src/infrastructure/api/generated/model/organizationMapMarkerDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Pet Rescue API
 * RESTful API for Pet Rescue and Adoption Platform
 * OpenAPI spec version: 1.0.0
 */

/**
 * Organization map marker
 */
export interface OrganizationMapMarkerDto {
  organizationId?: string;
  organizationCode?: string;
  name?: string;
  type?: string;
  status?: string;
  latitude?: number;
  longitude?: number;
  phone?: string;
  imageUrl?: string;
  wardName?: string;
  provinceName?: string;
}